+++ springboot/lms/src/main/java/com/ntuc/lms/repository/BorrowRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.ntuc.lms.model.Borrow;
 
-public interface BorrowRepository extends JpaRepository<Borrow, Long> {
+public interface BorrowRepository extends JpaRepository<Borrow, Integer> {
     List<Borrow> findByReturnedFalse();
 }
