+++ springboot/lms/src/main/java/com/ntuc/lms/repository/MemberRepository.java
@@ -4,4 +4,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.ntuc.lms.model.Member;
 
-public interface MemberRepository extends JpaRepository<Member, Long> {}
+public interface MemberRepository extends JpaRepository<Member, Integer> {}
